chore: update project layout
