+++ .gitignore
@@ -191,3 +191,6 @@ results/
 
 hack-browser-data
 !/cmd/hack-browser-data
+!/browingdata/history
+!/browingdata/history/history.go
+!/browingdata/history/history_test.go
