+++ browser/consts.go
@@ -1,0 +1,26 @@
+package browser
+
+import (
+	"os"
+)
+
+// home dir path for all platforms
+var homeDir, _ = os.UserHomeDir()
+
+const (
+	chromeName     = "Chrome"
+	chromeBetaName = "Chrome Beta"
+	chromiumName   = "Chromium"
+	edgeName       = "Microsoft Edge"
+	braveName      = "Brave"
+	operaName      = "Opera"
+	operaGXName    = "OperaGX"
+	vivaldiName    = "Vivaldi"
+	coccocName     = "CocCoc"
+	yandexName     = "Yandex"
+	firefoxName    = "Firefox"
+	speed360Name   = "360speed"
+	qqBrowserName  = "QQ"
+	dcBrowserName  = "DC"
+	sogouName      = "Sogou"
+)
